chore: remove superjson
