+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Removed
 - **BREAKING CHANGE**: Removed `DigestSecurityScheme`
+- **BREAKING CHANGE**: Removed superjson schema
 
 ### Fixed
 - Profile AST JSON Schema regenerated
