+++ src/interfaces/index.ts
@@ -1,3 +1,2 @@
 export * from './ast';
-export * from './superjson';
 export * from './providerjson';
