--- src/interfaces/superjson/index.ts
@@ -1,2 +1,0 @@
-export * from './superjson';
-export * from './superjson.utils';
